rm grant_eat process
