+++ .gitignore
@@ -305,6 +305,7 @@ philo/philo
 philo_bonus/philo_bonus
 .vscode/
 check/
+*.log
 
 # End of https://www.toptal.com/developers/gitignore/api/c,linux,python,macos,windows
