+++ philo/inc/actors/philo_actor.h
@@ -6,7 +6,7 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/06 16:28:12 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/30 15:58:00 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/30 16:13:57 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,6 @@ typedef struct s_philo_actor
 	size_t		no;
 	bool		has_l_fork;
 	bool		has_r_fork;
-	bool		can_eat;
 	bool		run_update;
 	t_timeval	last_update_time;
 }				t_philo_actor;
@@ -60,6 +59,7 @@ void			philo_on_start(t_ft_actor *self);
 bool			philo_on_receive(t_ft_actor *self, t_ft_msg *msg);
 void			philo_update(t_ft_actor *self);
 // methods
+void			_send_request_forks(t_philo_actor *self);
 void			_on_sync_start(t_philo_actor *self, t_ft_msg *msg);
 void			_common_update(t_philo_actor *self, long delta);
 void			_thinking(t_philo_actor *self);
