+++ philo/srcs/actors/philo/_thinking.c
@@ -6,27 +6,16 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/06 16:32:30 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/28 22:23:27 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/30 16:05:00 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo_actor.h"
 
-static void	_send_request_fork(t_ft_actor *fork, t_ft_actor *philo_ref)
-{
-	fork->tell(fork, msg_new(REQUEST_FORK, philo_ref));
-}
-
 void	_thinking(t_philo_actor *self)
 {
 	t_ft_msg	*msg;
 
-	if (self->can_eat)
-	{
-		_send_request_fork(self->l_fork, self->base);
-		_send_request_fork(self->r_fork, self->base);
-		self->can_eat = false;
-	}
 	if (self->has_l_fork && self->has_r_fork)
 	{
 		msg = msg_new(MONITOR_EATING, self->base);
