+++ philo/srcs/actors/philo/philo_on_start.c
@@ -6,7 +6,7 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/06 16:32:30 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/24 22:27:43 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/30 16:05:07 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,5 @@ void	philo_on_start(t_ft_actor *self)
 	philo->eat_count = 0;
 	philo->has_r_fork = false;
 	philo->has_l_fork = false;
-	philo->can_eat = false;
 	parent->tell(parent, msg_new(ACTOR_START_DONE, self));
 }
